+++ Runtime/Android/AndroidJavaProxies/ResultString.cs
@@ -1,5 +1,6 @@
 using Chartboost.Core.Android.Utilities;
 using Chartboost.Core.Utilities;
+using UnityEngine;
 using UnityEngine.Scripting;
 
 namespace Chartboost.Core.Android.AndroidJavaProxies
@@ -18,10 +19,15 @@ namespace Chartboost.Core.Android.AndroidJavaProxies
         /// <param name="result">String value.</param>
         [Preserve]
         // ReSharper disable once InconsistentNaming
-        private void onResult(string result)
-        {
-            MainThreadDispatcher.Post(o => _complete(result));
-        }
+        private void onResult(string result) => MainThreadDispatcher.Post(o => _complete(result));
+        
+        /// <summary>
+        /// Posts a null result from the native layer.
+        /// </summary>
+        /// <param name="result">Null value.</param>
+        [Preserve]
+        // ReSharper disable once InconsistentNaming
+        private void onResult(AndroidJavaObject result) => MainThreadDispatcher.Post(o => _complete(null));
     }
     #nullable disable
 }