+++ Runtime/Android/Environment/PublisherMetadata.cs
@@ -1,4 +1,5 @@
 using System;
+using Chartboost.Core.Android.AndroidJavaProxies;
 using Chartboost.Core.Android.Utilities;
 using Chartboost.Core.Environment;
 using UnityEngine;
@@ -18,6 +19,20 @@ namespace Chartboost.Core.Android.Environment
         /// <inheritdoc cref="BaseAndroidEnvironment.EnvironmentBridge"/>
         protected override Func<AndroidJavaClass> EnvironmentBridge => null!;
 
+        public PublisherMetadata()
+        {
+            using var sdk = AndroidUtils.NativeSDK();
+            using var publisherMetadata = sdk.CallStatic<AndroidJavaObject>(EnvironmentProperty);
+            publisherMetadata.Call(AndroidConstants.AddObserver, new PublisherMetadataObserver(this));
+        }
+
+        public event Action? IsUserUnderageChanged;
+        public event Action? PublisherSessionIdentifierChanged;
+        public event Action? PublisherAppIdentifierChanged;
+        public event Action? FrameworkNameChanged;
+        public event Action? FrameworkVersionChanged;
+        public event Action? PlayerIdentifierChanged;
+
         /// <inheritdoc cref="IPublisherMetadata.SetIsUserUnderage"/>
         public void SetIsUserUnderage(bool isUserUnderage) => SetProperty(AndroidConstants.SetPropertyIsUserUnderAge, isUserUnderage);
         
@@ -35,6 +50,32 @@ namespace Chartboost.Core.Android.Environment
         
         /// <inheritdoc cref="IPublisherMetadata.SetPlayerIdentifier"/>
         public void SetPlayerIdentifier(string? playerIdentifier) => SetProperty(AndroidConstants.SetPropertyPlayerIdentifier, playerIdentifier);
+
+        ~PublisherMetadata()
+        {
+            AndroidJNI.AttachCurrentThread();
+            using var consentManagementPlatform = AndroidUtils.ConsentManagementPlatform();
+            consentManagementPlatform.Call(AndroidConstants.RemoveObserver, PublisherMetadataObserver.Instance);
+            AndroidJNI.DetachCurrentThread();
+        }
+
+        internal void OnIsUserUnderageChanged()
+            => IsUserUnderageChanged?.Invoke();
+
+        internal void OnPublisherSessionIdentifierChanged()
+            => PublisherSessionIdentifierChanged?.Invoke();
+
+        internal void OnPublisherAppIdentifierChanged() 
+            => PublisherAppIdentifierChanged?.Invoke();
+
+        internal void OnFrameworkNameChanged() 
+            => FrameworkNameChanged?.Invoke();
+
+        internal void OnFrameworkVersionChanged() 
+            => FrameworkVersionChanged?.Invoke();
+
+        internal void OnPlayerIdentifierChanged() 
+            => PlayerIdentifierChanged?.Invoke();
     }
     #nullable disable
 }