+++ Runtime/Core/Modules/TestUnityModuleFailure.cs
@@ -1,6 +1,8 @@
+using System;
 using System.Threading.Tasks;
 using Chartboost.Core.Error;
 using Chartboost.Core.Initialization;
+using Newtonsoft.Json;
 
 namespace Chartboost.Core.Modules
 {
@@ -11,9 +13,9 @@ namespace Chartboost.Core.Modules
         
         public static readonly ChartboostCoreError Error = new ChartboostCoreError(500, "Module failed to initialize", "This module always fails to initialize", "nothing...");
 
-        protected override Task<ChartboostCoreError?> Initialize()
+        protected override Task<ChartboostCoreError?> Initialize(ModuleInitializationConfiguration configuration)
         {
-            ChartboostCoreLogger.Log($"[{ModuleId}/{ModuleVersion}] Attempting Initialization. Always fail.");
+            ChartboostCoreLogger.Log($"[{ModuleId}/{ModuleVersion}] Attempting Initialization. Always fail and Config: {JsonConvert.SerializeObject(configuration)}");
             return Task.FromResult<ChartboostCoreError?>(Error);
         }
     }