+++ Runtime/Core/Modules/TestUnityModuleSuccess.cs
@@ -2,6 +2,7 @@ using System;
 using System.Threading.Tasks;
 using Chartboost.Core.Error;
 using Chartboost.Core.Initialization;
+using Newtonsoft.Json;
 
 namespace Chartboost.Core.Modules
 {
@@ -10,9 +11,9 @@ namespace Chartboost.Core.Modules
         public override string ModuleId => "Unity Test Module Success";
         public override string ModuleVersion => "0.0.1";
 
-        protected override Task<ChartboostCoreError?> Initialize()
+        protected override Task<ChartboostCoreError?> Initialize(ModuleInitializationConfiguration configuration)
         {
-            ChartboostCoreLogger.Log($"[{ModuleId}/{ModuleVersion}] Attempting Initialization. Always succeed");
+            ChartboostCoreLogger.Log($"[{ModuleId}/{ModuleVersion}] Attempting Initialization. Always succeed, and Config {JsonConvert.SerializeObject(configuration)}");
             return Task.FromResult<ChartboostCoreError?>(null);
         }
     }