+++ Runtime/iOS/Plugins/iOS/Observers/CBCUnityObserver.h
@@ -1,6 +1,6 @@
 #import "CBCUnityUtilities.h"
 
-@interface CBCUnityObserver : NSObject<CBCInitializableModuleObserver, CBCConsentObserver>
+@interface CBCUnityObserver : NSObject<CBCInitializableModuleObserver, CBCConsentObserver, CBCPublisherMetadataObserver>
 
 + (instancetype) sharedObserver;
 - (NSMutableDictionary*) initializableModules;
@@ -10,9 +10,16 @@
 - (void) clearInitializableModules;
 - (void) storeModule:(id<CBCInitializableModule>)nativeModule;
 
-@property ChartboostCoreOnModuleInitializationResultCallback onModuleInitializationCompleted;
-@property ChartboostCoreOnConsentStatusChangeCallback onConsentStatusChange;
-@property ChartboostCoreOnConsentChangeCallback onConsentChangeForStandard;
+#pragma mark CBCInitializableModuleObserver
+@property ChartboostCoreOnModuleInitializationResult onModuleInitializationCompleted;
+
+#pragma mark CBCConsentObserver
+@property ChartboostCoreOnEnumStatusChange onConsentStatusChange;
+@property ChartboostCoreOnConsentChangeForStandard onConsentChangeForStandard;
 @property ChartboostCoreAction onConsentReady;
+
+#pragma mark CBCPublisherMetadataObserver
+
+@property ChartboostCoreOnEnumStatusChange onPublisherMetadataPropertyChange;
 
 @end